requirements.txt
discord.py
aichttp
websockets
chardet